auth-service/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
httpx==0.27.2
prometheus-fastapi-instrumentator==7.0.2
python-json-logger==2.0.7
pytest==8.3.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
